pyproject.toml
[project]
name = "ibauth"
version = "0.1.2"
description = "Interactive Brokers OAuth"
authors = [
  {name = "Andrew Collier"}
]
maintainers = [
  {name = "Andrew Collier"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.2",
    "httpx>=0.28.1",
    "pydantic>=2.11.10",
    "pydantic-extra-types>=2.11.0",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "tenacity>=9.0.0",
]

[dependency-groups]
dev = [
    "bump2version>=1.0.1",
    "pre-commit>=4.1.0",
    "ty>=0.0.14",
    "ruff>=0.13.3",
    "pytest-asyncio>=1.1.0",
    "pictex>=1.5.0",
]
test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.1.1",
    "pytest-timeout>=2.4.0",
]

[build-system]
requires = ["setuptools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--timeout=10"]
markers = [
  "integration: Tests that require a running ibproxy.",
  "no_patch_connect: Tests that require a _connect() method.",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "F"]

[tool.uv]
default-groups = ["dev", "test"]

[tool.coverage.report]
include = ["src/*"]
show_missing = true
skip_covered = false
precision = 1

[tool.ty.environment]
python = "./.venv"

[project.scripts]
ibauth = "ibauth:main"
